feat: split直接拆分
